Core/Units: Reduce differences between branches part 1 - object class
