+++ src/server/game/Entities/Unit/Unit.h
@@ -1830,7 +1830,7 @@ class TC_GAME_API Unit : public WorldObject
     protected:
         explicit Unit (bool isWorldObject);
 
-        void BuildValuesUpdate(uint8 updatetype, ByteBuffer* data, Player* target) const override;
+        void BuildValuesUpdate(uint8 updatetype, ByteBuffer* data, Player const* target) const override;
 
         void _UpdateSpells(uint32 time);
         void _DeleteRemovedAuras();
